ECS dependencies sorted out and moved Dipole and Cooling Ligth back to laser module
