+++ examples/molasses_1d.rs
@@ -4,8 +4,8 @@ use lib::atom::{Atom, AtomicTransition, Force, Mass, Position, Velocity};
 use lib::ecs;
 use lib::initiate::NewlyCreated;
 use lib::integrator::Timestep;
+use lib::laser::cooling::CoolingLight;
 use lib::laser::gaussian::GaussianBeam;
-use lib::laser_cooling::cooling::CoolingLight;
 use lib::laser_cooling::photons_scattered::ActualPhotonsScatteredVector;
 use lib::output::file;
 use lib::output::file::Text;
