+++ src/dipole/transition_switcher.rs
@@ -4,10 +4,10 @@ use crate::atom::{Atom, AtomicTransition, Kind};
 use crate::constant::C;
 use crate::destructor::ToBeDestroyed;
 use crate::dipole::atom::AtomicDipoleTransition;
-use crate::dipole::dipole_beam::DipoleLight;
 use crate::integrator::Timestep;
+use crate::laser::cooling::CoolingLight;
+use crate::laser::dipole_beam::DipoleLight;
 use crate::laser::gaussian::GaussianBeam;
-use crate::laser_cooling::cooling::CoolingLight;
 use specs::ReadExpect;
 use specs::{Entities, Join, LazyUpdate, Read, ReadStorage, System, WriteStorage};
 extern crate nalgebra;
