+++ src/integration_tests/rate_equation.rs
@@ -8,8 +8,8 @@ pub mod tests {
     use crate::ecs;
     use crate::initiate::NewlyCreated;
     use crate::integrator::Timestep;
+    use crate::laser::cooling::{CoolingLight, CoolingLightIndex};
     use crate::laser::gaussian::GaussianBeam;
-    use crate::laser_cooling::cooling::{CoolingLight, CoolingLightIndex};
     use crate::laser_cooling::photons_scattered::TotalPhotonsScattered;
     extern crate nalgebra;
     use nalgebra::Vector3;
