+++ src/laser/cooling.rs
