+++ src/laser/dipole_beam.rs
