+++ src/laser/mod.rs
@@ -1,5 +1,7 @@
 //! Calculation and initialization of optical forces and quantities exerted on the atoms
 
+pub mod cooling;
+pub mod dipole_beam;
 pub mod frame;
 pub mod gaussian;
 pub mod intensity;
@@ -61,8 +63,28 @@ impl<'a> System<'a> for AttachLaserComponentsToNewlyCreatedAtomsSystem {
 pub fn add_systems_to_dispatch(builder: &mut DispatcherBuilder<'static, 'static>, deps: &[&str]) {
 	builder.add(
 		AttachLaserComponentsToNewlyCreatedAtomsSystem,
-		"attach_atom_laser_components",
+		"attach_laser_components",
 		deps,
+	);
+	builder.add(
+		cooling::AttachIndexToCoolingLightSystem,
+		"attach_cooling_index",
+		deps,
+	);
+	builder.add(
+		cooling::IndexCoolingLightsSystem,
+		"index_cooling_lights",
+		deps,
+	);
+	builder.add(
+		dipole_beam::AttachIndexToDipoleLightSystem,
+		"attach_dipole_index",
+		deps,
+	);
+	builder.add(
+		dipole_beam::IndexDipoleLightsSystem,
+		"index_dipole_lights",
+		&["attach_dipole_index"],
 	);
 	builder.add(
 		sampler::InitialiseLaserSamplerMasksSystem,
@@ -91,4 +113,6 @@ pub fn register_components(world: &mut World) {
 	world.register::<gaussian::GaussianBeam>();
 	world.register::<gaussian::CircularMask>();
 	world.register::<gaussian::GaussianReferenceFrame>();
+	world.register::<dipole_beam::DipoleLight>();
+	world.register::<dipole_beam::DipoleLightIndex>();
 }
